requirements.txt
ipykernel==6.25.2
boto3==1.28.67
sqlparse==0.4.4
plotly==5.17.0
dash==2.14.0
dash-ag-grid==2.4.0
dash-bootstrap-components==1.5.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
pyarrow==13.0.0
pandasql==0.7.3
pivottablejs==0.9.0
ipython-sql==0.5.0
jupyter-resource-usage==1.0.1
pygwalker>=0.3.9,<=1.0.0